accept part of pull_request events experimentally
rationale: if you have heard that linus greatly values diffstat
and shortlog, well for me i deeply value range-diffs across patch
revisions... and the way github does it just really doesn't it
cut it for me.  like no sane person would look at the suggested
changes and go "yeah, that makes sense to me" -- it's a muddle of
porridge is what it is.  (乱成一锅粥 is a proverb that i never
knew how to translate, but something like that i guess... :/)